+++ tardis/spindletorch/unet/network.py
@@ -62,6 +62,7 @@ class UNet(nn.Module):
                  no_groups=8,
                  prediction=False):
         super(UNet, self).__init__()
+        self.prediction = prediction
 
         patch_sizes = [patch_size]
         for _ in range(no_conv_layer):
@@ -69,7 +70,7 @@ class UNet(nn.Module):
             patch_sizes.append(patch_size)
         patch_sizes = list(reversed(patch_sizes))[2:]
 
-        self.prediction = prediction
+        """ Encoder """
         self.encoder = build_encoder(in_ch=in_channels,
                                      conv_layers=no_conv_layer,
                                      conv_layer_multiplayer=conv_layer_multiplayer,
@@ -81,6 +82,7 @@ class UNet(nn.Module):
                                      pool_kernel=pool_kernel,
                                      conv_module=DoubleConvolution)
 
+        """ Decoder """
         self.decoder = build_decoder(conv_layers=no_conv_layer,
                                      conv_layer_multiplayer=conv_layer_multiplayer,
                                      components=layer_components,
@@ -88,6 +90,7 @@ class UNet(nn.Module):
                                      no_groups=no_groups,
                                      deconv_module='CNN')
 
+        """ Final Layer """
         if '3' in layer_components:
             self.final_conv_layer = nn.Conv3d(in_channels=conv_layer_multiplayer,
                                               out_channels=out_channels,
@@ -97,6 +100,7 @@ class UNet(nn.Module):
                                               out_channels=out_channels,
                                               kernel_size=1)
 
+        """ Prediction """
         if sigmoid:
             self.activation = nn.Sigmoid()
         else:
@@ -107,18 +111,20 @@ class UNet(nn.Module):
         encoder_features = []
 
         """ Encoder """
-        for encoder in self.encoder:
+        for i, encoder in enumerate(self.encoder):
             x = encoder(x)
-            encoder_features.insert(0, x)
 
-        encoder_features = encoder_features[1:]
+            if (len(self.encoder) - 1) != i:
+                encoder_features.insert(0, x)
 
         """ Decoder """
         for decoder, encoder_features in zip(self.decoder, encoder_features):
             x = decoder(encoder_features, x)
 
+        """ Final Layer """
         x = self.final_conv_layer(x)
 
+        """ Prediction """
         if self.prediction:
             x = self.activation(x)
 
@@ -148,6 +154,7 @@ class ResUNet(nn.Module):
                  no_groups=8,
                  prediction=False):
         super(ResUNet, self).__init__()
+        self.prediction = prediction
 
         patch_sizes = [patch_size]
         for _ in range(no_conv_layer):
@@ -156,7 +163,7 @@ class ResUNet(nn.Module):
 
         patch_sizes = list(reversed(patch_sizes))[2:]
 
-        self.prediction = prediction
+        """ Encoder """
         self.encoder = build_encoder(in_ch=in_channels,
                                      conv_layers=no_conv_layer,
                                      conv_layer_multiplayer=conv_layer_multiplayer,
@@ -176,6 +183,7 @@ class ResUNet(nn.Module):
                                      no_groups=no_groups,
                                      deconv_module='RCNN')
 
+        """ Final Layer """
         if '3' in layer_components:
             self.final_conv_layer = nn.Conv3d(in_channels=conv_layer_multiplayer,
                                               out_channels=out_channels,
@@ -185,6 +193,7 @@ class ResUNet(nn.Module):
                                               out_channels=out_channels,
                                               kernel_size=1)
 
+        """ Prediction """
         if sigmoid:
             self.activation = nn.Sigmoid()
         else:
@@ -195,18 +204,20 @@ class ResUNet(nn.Module):
         encoder_features = []
 
         """ Encoder """
-        for encoder in self.encoder:
+        for i, encoder in enumerate(self.encoder):
             x = encoder(x)
-            encoder_features.insert(0, x)
 
-        encoder_features = encoder_features[1:]
+            if (len(self.encoder) - 1) != i:
+                encoder_features.insert(0, x)
 
         """ Decoder """
         for decoder, encoder_features in zip(self.decoder, encoder_features):
             x = decoder(encoder_features, x)
 
+        """ Final Layer """
         x = self.final_conv_layer(x)
 
+        """ Prediction """
         if self.prediction:
             x = self.activation(x)
 
@@ -228,7 +239,7 @@ class UNet3Plus(nn.Module):
                  patch_size=64,
                  no_conv_layer=4,
                  conv_layer_multiplayer=64,
-                 layer_components="cgl",
+                 layer_components="3cgl",
                  no_groups=8,
                  conv_kernel=3,
                  pool_kernel=2,
@@ -236,6 +247,7 @@ class UNet3Plus(nn.Module):
                  classifies=False,
                  prediction=False):
         super(UNet3Plus, self).__init__()
+        self.prediction = prediction
 
         patch_sizes = [patch_size]
         for _ in range(no_conv_layer):
@@ -247,7 +259,7 @@ class UNet3Plus(nn.Module):
         feature_map = number_of_features_per_level(conv_layer_multiplayer,
                                                    no_conv_layer)
 
-        self.prediction = prediction
+        """ Encoder """
         self.encoder = build_encoder(in_ch=in_channels,
                                      conv_layers=no_conv_layer,
                                      conv_layer_multiplayer=conv_layer_multiplayer,
@@ -258,6 +270,7 @@ class UNet3Plus(nn.Module):
                                      pool_kernel=pool_kernel,
                                      conv_module=DoubleConvolution)
 
+        """ UNet3Plus classifier """
         if classifies:
             if '3' in layer_components:
                 self.cls = nn.Sequential(nn.Dropout(p=0.5),
@@ -276,6 +289,7 @@ class UNet3Plus(nn.Module):
         else:
             self.cls = None
 
+        """ Decoder """
         self.decoder = build_decoder(conv_layers=no_conv_layer,
                                      conv_layer_multiplayer=conv_layer_multiplayer,
                                      components=layer_components,
@@ -283,6 +297,7 @@ class UNet3Plus(nn.Module):
                                      no_groups=no_groups,
                                      deconv_module='unet3plus')
 
+        """ Final Layer """
         if '3' in layer_components:
             self.final_conv_layer = nn.Conv3d(in_channels=conv_layer_multiplayer,
                                               out_channels=out_channels,
@@ -292,6 +307,7 @@ class UNet3Plus(nn.Module):
                                               out_channels=out_channels,
                                               kernel_size=1)
 
+        """ Prediction """
         if sigmoid:
             self.activation = nn.Sigmoid()
         else:
@@ -338,7 +354,7 @@ class UNet3Plus(nn.Module):
             decoder_features.insert(0, x)
             encoder_features = encoder_features[1:]
 
-        # for i, decoder in enumerate(decoder_features):
+        """ Final Layer / Prediction"""
         if self.cls is not None:
             x = self.final_conv_layer(x)
             x = self.dotProduct(x, x_cls_max)
@@ -354,3 +370,134 @@ class UNet3Plus(nn.Module):
                 x = self.activation(x)
 
             return x
+
+
+class Big_UNet(nn.Module):
+    """
+    New Unet model combining Unet and Unet3Plus
+    Model shares encoder path which is splitted for decoding patch Unet and Unet3Plus
+    style. Final layers from each are summed and sigmoid 
+
+    Args:
+        in_channels: Number of input channels for first convolution
+        out_channels: Number of output channels for last deconvolution
+        sigmoid: If True, use nn.Sigmoid or nn.Softmax if False. Use True if
+            nn.BCELoss is used as loss function for (two-class segmentation)
+        no_conv_layer: Number of convolution and deconvolution steps. Number of
+            input channels for convolution is calculated as a linear progression.
+            E.g. [64, 128, 256, 512]
+        conv_layer_multiplayer: Feature output of first layer
+        conv_kernel: Kernel size for the convolution
+        padding: Padding size for convolution
+        maxpool_kernel: kernel size for max_pooling
+        patch_size: Image patch size used for calculation network structure
+        layer_components: Convolution module used for build network
+        no_groups: Number of group for nn.GroupNorm
+        prediction: If True, prediction mode is on
+    """
+
+    def __init__(self,
+                 in_channels=1,
+                 out_channels=1,
+                 sigmoid=True,
+                 no_conv_layer=5,
+                 conv_layer_multiplayer=64,
+                 conv_kernel=3,
+                 padding=1,
+                 pool_kernel=2,
+                 patch_size=64,
+                 layer_components="3gcl",
+                 no_groups=8,
+                 prediction=False):
+        super(Big_UNet, self).__init__()
+        self.prediction = prediction
+
+        patch_sizes = [patch_size]
+        for _ in range(no_conv_layer):
+            patch_size = int(patch_size / 2)
+            patch_sizes.append(patch_size)
+        patch_sizes = list(reversed(patch_sizes))[2:]
+
+        """ Encoder """
+        self.encoder = build_encoder(in_ch=in_channels,
+                                     conv_layers=no_conv_layer,
+                                     conv_layer_multiplayer=conv_layer_multiplayer,
+                                     conv_kernel=conv_kernel,
+                                     padding=padding,
+                                     no_groups=no_groups,
+                                     components=layer_components,
+                                     pool_kernel=pool_kernel,
+                                     conv_module=DoubleConvolution)
+
+        """ Decoder """
+        self.decoder_unet = build_decoder(conv_layers=no_conv_layer,
+                                          conv_layer_multiplayer=conv_layer_multiplayer,
+                                          components=layer_components,
+                                          sizes=patch_sizes,
+                                          no_groups=no_groups,
+                                          deconv_module='CNN')
+        self.decoder_3plus = build_decoder(conv_layers=no_conv_layer,
+                                           conv_layer_multiplayer=conv_layer_multiplayer,
+                                           components=layer_components,
+                                           sizes=patch_sizes,
+                                           no_groups=no_groups,
+                                           deconv_module='unet3plus')
+
+        """ Final Layer """
+        if '3' in layer_components:
+            self.final_conv_layer_unet = nn.Conv3d(in_channels=conv_layer_multiplayer,
+                                                   out_channels=out_channels,
+                                                   kernel_size=1)
+            self.final_conv_layer_3plus = nn.Conv3d(in_channels=conv_layer_multiplayer,
+                                                    out_channels=out_channels,
+                                                    kernel_size=1)
+        elif '2' in layer_components:
+            self.final_conv_layer_unet = nn.Conv2d(in_channels=conv_layer_multiplayer,
+                                                   out_channels=out_channels,
+                                                   kernel_size=1)
+            self.final_conv_layer_3plus = nn.Conv2d(in_channels=conv_layer_multiplayer,
+                                                    out_channels=out_channels,
+                                                    kernel_size=1)
+
+        """ Prediction """
+        if sigmoid:
+            self.activation = nn.Sigmoid()
+        else:
+            self.activation = nn.Softmax(dim=1)
+
+    def forward(self,
+                x: torch.Tensor):
+        encoder_features = []
+
+        """ Encoder """
+        for i, encoder in enumerate(self.encoder):
+            x = encoder(x)
+
+            if (len(self.encoder) - 1) != i:
+                encoder_features.insert(0, x)
+
+        """ Decoder UNet """
+        x_unet = x
+        for decoder, features in zip(self.decoder_unet, encoder_features):
+            x_unet = decoder(features, x_unet)
+
+        """ Decoder UNet_3Plus """
+        x_3plus = x
+        for decoder in self.decoder_3plus:
+            decoder_features = [x_3plus]
+
+            x_3plus = decoder(x=x_3plus,
+                              encoder_features=encoder_features,
+                              decoder_features=decoder_features[2:])
+
+            # add/remove layer at each iter
+            decoder_features.insert(0, x_3plus)
+            encoder_features = encoder_features[1:]
+
+        """ Final Layer/Prediction """
+        if self.prediction:
+            return self.activation(torch.add(self.final_conv_layer_unet(x_unet),
+                                             self.final_conv_layer_3plus(x_3plus)))
+        else:
+            return torch.add(self.final_conv_layer_unet(x_unet),
+                             self.final_conv_layer_3plus(x_3plus))