+++ tests/test_spindetorch_nn.py
@@ -116,6 +116,40 @@ class TestNetwork3D:
                              no_groups=8,
                              prediction=False)
 
+    def big_unet_2d(self,
+                    image_size: int,
+                    layer_no: int,
+                    conv_multiplayer: int,
+                    dropout: Optional[float] = None):
+        return build_network(network_type='big_unet',
+                             classification=False,
+                             in_channel=1,
+                             out_channel=1,
+                             img_size=image_size,
+                             dropout=dropout,
+                             no_conv_layers=layer_no,
+                             conv_multiplayer=conv_multiplayer,
+                             layer_components='2gcl',
+                             no_groups=8,
+                             prediction=False)
+
+    def big_unet_3d(self,
+                    image_size: int,
+                    layer_no: int,
+                    conv_multiplayer: int,
+                    dropout: Optional[float] = None):
+        return build_network(network_type='big_unet',
+                             classification=False,
+                             in_channel=1,
+                             out_channel=1,
+                             img_size=image_size,
+                             dropout=dropout,
+                             no_conv_layers=layer_no,
+                             conv_multiplayer=conv_multiplayer,
+                             layer_components='3gcl',
+                             no_groups=8,
+                             prediction=False)
+
     def test_unet3d(self):
         for i in self.image_sizes:
             for j in self.conv_multiplayers:
@@ -172,6 +206,20 @@ class TestNetwork3D:
                 with torch.no_grad():
                     input = nn(input)
 
+    def test_unet3plus_2d(self):
+        for i in self.image_sizes:
+            for j in self.conv_multiplayers:
+                # Batch x Channel x Z x Y x X
+                input = torch.rand((1, 1, i, i))
+
+                nn = self.unet3plus_2d(image_size=i,
+                                       layer_no=5,
+                                       conv_multiplayer=j,
+                                       dropout=None)
+
+                with torch.no_grad():
+                    input = nn(input)
+
     def test_unet3plus_3d(self):
         for i in self.image_sizes:
             for j in self.conv_multiplayers:
@@ -186,16 +234,30 @@ class TestNetwork3D:
                 with torch.no_grad():
                     input = nn(input)
 
-    def test_unet3plus_2d(self):
+    def test_big_unet_2d(self):
         for i in self.image_sizes:
             for j in self.conv_multiplayers:
                 # Batch x Channel x Z x Y x X
                 input = torch.rand((1, 1, i, i))
 
-                nn = self.unet3plus_2d(image_size=i,
-                                       layer_no=5,
-                                       conv_multiplayer=j,
-                                       dropout=None)
+                nn = self.big_unet_2d(image_size=i,
+                                      layer_no=5,
+                                      conv_multiplayer=j,
+                                      dropout=None)
+
+                with torch.no_grad():
+                    input = nn(input)
+
+    def test_big_unet_3d(self):
+        for i in self.image_sizes:
+            for j in self.conv_multiplayers:
+                # Batch x Channel x Z x Y x X
+                input = torch.rand((1, 1, i, i, i))
+
+                nn = self.big_unet_3d(image_size=i,
+                                      layer_no=5,
+                                      conv_multiplayer=j,
+                                      dropout=None)
 
                 with torch.no_grad():
                     input = nn(input)